refactor: move (a lot) of code around
