+++ apps/antalmanac/src/stores/TabStore.ts
@@ -18,5 +18,3 @@ export const useTabStore = create<TabStore>((set) => {
         },
     };
 });
-
-export default useTabStore;
